pyproject.toml
[project]
name = "huggingface-model-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.28.1",
    "pyqt6>=6.8.1",
    "python-dotenv>=1.0.1",
    "torch>=2.6.0",
    "transformers>=4.48.3",
]
